fix getDerivedStateFromProps for react v16.4+
From v16.4 it is called everytime component is rerendered, does not
matter if it is caused by props or state update. Previously it was only
called when props are changed. That is why we have to
check if the props are really changed by keeping the old props in state
and doing a comparison.

Unfortunately this is not the best fix we can do, it is better to
refactor to code not to use getDerivedStateFromProps if possible.
(See INSPIR-1537)